+++ tools/build_langserver/langserver/utils.go
@@ -12,6 +12,7 @@ import (
 	"regexp"
 	"strings"
 
+	"parse/asp"
 	"tools/build_langserver/lsp"
 )
 
@@ -215,11 +216,6 @@ func ExtractLiteral(str string) string {
 	return ""
 }
 
-//
-//func ExtractFuncCall(str string) (string, []string) {
-//
-//}
-
 func mustMatch(pattern string, str string) bool {
 	re := regexp.MustCompile(pattern)
 	matched := re.FindString(str)
@@ -232,4 +228,44 @@ func mustMatch(pattern string, str string) bool {
 // isEmpty checks if the hovered line is empty
 func isEmpty(lineContent string, pos lsp.Position) bool {
 	return len(lineContent) < pos.Character || strings.TrimSpace(lineContent[:pos.Character]) == ""
+}
+
+// withInRange checks if the input asp.Position from lsp is within the range of the Expression
+func withInRange(exprPos asp.Position, exprEndPos asp.Position, pos lsp.Position) bool {
+	withInLineRange := pos.Line >= exprPos.Line-1 &&
+		pos.Line <= exprEndPos.Line-1
+
+	withInColRange := pos.Character >= exprPos.Column-1 &&
+		pos.Character <= exprEndPos.Column-1
+
+	onTheSameLine := pos.Line == exprEndPos.Line-1 &&
+		pos.Line == exprPos.Line-1
+
+	if !withInLineRange || (onTheSameLine && !withInColRange) {
+		return false
+	}
+
+	if pos.Line == exprPos.Line-1 {
+		return pos.Character >= exprPos.Column-1
+	}
+
+	if pos.Line == exprEndPos.Line-1 {
+		return pos.Character <= exprEndPos.Column-1
+	}
+
+	return true
+}
+
+func withInRangeLSP(targetPos lsp.Position, targetEndPos lsp.Position, pos lsp.Position) bool {
+	start := lspPositionToAsp(targetPos)
+	end := lspPositionToAsp(targetEndPos)
+
+	return withInRange(start, end, pos)
+}
+
+func lspPositionToAsp(pos lsp.Position) asp.Position {
+	return asp.Position{
+		Line:   pos.Line + 1,
+		Column: pos.Character + 1,
+	}
 }
